Implement q_shuffle using Fisher-Yates algorithm
Implemented `q_shuffle` function based on the modern method of the Fisher-Yates
shuffle algorithm as described in its Wikipedia article. This approach involves
using a pointer `back` to refer to the last element in the list. A random number
in the range 0 to (size-1) is generated to determine the node to swap with.
In each iteration, the `cur` and `back` nodes are swapped, and then `back` is
decremented to progressively reduce the range of random numbers, thus
completing the shuffling process.